提取公共Head
