+++ pages/_app.js
@@ -18,6 +18,7 @@ import useAdjustStyle from '@/hooks/useAdjustStyle'
 // 各种扩展插件 这个要阻塞引入
 import ExternalPlugins from '@/components/ExternalPlugins'
 import { THEME } from '@/blog.config'
+import GlobalHead from '@/components/GlobalHead'
 
 const MyApp = ({ Component, pageProps }) => {
   // 一些可能出现 bug 的样式，可以统一放入该钩子进行调整
@@ -28,6 +29,7 @@ const MyApp = ({ Component, pageProps }) => {
     return getQueryParam(route.asPath, 'theme') || THEME
   }, [route])
 
+  // 整体布局
   const GLayout = useCallback(
     props => {
       // 根据页面路径加载不同Layout文件
@@ -40,6 +42,7 @@ const MyApp = ({ Component, pageProps }) => {
   return (
     <GlobalContextProvider {...pageProps}>
       <GLayout {...pageProps}>
+        <GlobalHead {...pageProps}/>
         <Component {...pageProps} />
       </GLayout>
       <ExternalPlugins {...pageProps} />
