+++ pages/archive/index.js
@@ -1,6 +1,5 @@
 import { getGlobalData } from '@/lib/notion/getNotionData'
 import { useEffect } from 'react'
-import { useGlobal } from '@/lib/global'
 import BLOG from '@/blog.config'
 import { useRouter } from 'next/router'
 import { getLayoutByTheme } from '@/themes/theme'
@@ -9,9 +8,6 @@ import { formatDateFmt } from '@/lib/formatDate'
 import { siteConfig } from '@/lib/config'
 
 const ArchiveIndex = props => {
-  const { siteInfo } = props
-  const { locale } = useGlobal()
-
   // 根据页面路径加载不同Layout文件
   const Layout = getLayoutByTheme({ theme: siteConfig('THEME'), router: useRouter() })
 
@@ -28,16 +24,6 @@ const ArchiveIndex = props => {
       }
     }
   }, [])
-
-  const meta = {
-    title: `${locale.NAV.ARCHIVE} | ${siteConfig('TITLE')}`,
-    description: siteConfig('DESCRIPTION'),
-    image: siteInfo?.pageCover,
-    slug: 'archive',
-    type: 'website'
-  }
-
-  props = { ...props, meta }
 
   return <Layout {...props} />
 }
