+++ pages/category/[category]/index.js
@@ -1,6 +1,5 @@
 import { getGlobalData } from '@/lib/notion/getNotionData'
 import React from 'react'
-import { useGlobal } from '@/lib/global'
 import BLOG from '@/blog.config'
 import { useRouter } from 'next/router'
 import { getLayoutByTheme } from '@/themes/theme'
@@ -12,23 +11,8 @@ import { siteConfig } from '@/lib/config'
  * @returns
  */
 export default function Category(props) {
-  const { siteInfo } = props
-  const { locale } = useGlobal()
-
   // 根据页面路径加载不同Layout文件
   const Layout = getLayoutByTheme({ theme: siteConfig('THEME'), router: useRouter() })
-
-  const meta = {
-    title: `${props.category} | ${locale.COMMON.CATEGORY} | ${
-      siteConfig('TITLE') || ''
-    }`,
-    description: siteConfig('DESCRIPTION'),
-    slug: 'category/' + props.category,
-    image: siteInfo?.pageCover,
-    type: 'website'
-  }
-
-  props = { ...props, meta }
 
   return <Layout {...props} />
 }
