+++ pages/search/[keyword]/page/[page].js
@@ -1,5 +1,4 @@
 import { getGlobalData } from '@/lib/notion/getNotionData'
-import { useGlobal } from '@/lib/global'
 import { getDataFromCache } from '@/lib/cache/cache_manager'
 import BLOG from '@/blog.config'
 import { useRouter } from 'next/router'
@@ -7,21 +6,10 @@ import { getLayoutByTheme } from '@/themes/theme'
 import { siteConfig } from '@/lib/config'
 
 const Index = props => {
-  const { keyword, siteInfo } = props
-  const { locale } = useGlobal()
-
+  const { keyword } = props
   // 根据页面路径加载不同Layout文件
   const Layout = getLayoutByTheme({ theme: siteConfig('THEME'), router: useRouter() })
-
-  const meta = {
-    title: `${keyword || ''}${keyword ? ' | ' : ''}${locale.NAV.SEARCH} | ${siteConfig('TITLE')}`,
-    description: siteConfig('TITLE'),
-    image: siteInfo?.pageCover,
-    slug: 'search/' + (keyword || ''),
-    type: 'website'
-  }
-
-  props = { ...props, meta, currentSearch: keyword }
+  props = { ...props, currentSearch: keyword }
 
   return <Layout {...props} />
 }
