+++ pages/search/index.js
@@ -1,5 +1,4 @@
 import { getGlobalData } from '@/lib/notion/getNotionData'
-import { useGlobal } from '@/lib/global'
 import { useRouter } from 'next/router'
 import BLOG from '@/blog.config'
 import { getLayoutByTheme } from '@/themes/theme'
@@ -11,14 +10,13 @@ import { siteConfig } from '@/lib/config'
  * @returns
  */
 const Search = props => {
-  const { posts, siteInfo } = props
-  const { locale } = useGlobal()
+  const { posts } = props
 
   // 根据页面路径加载不同Layout文件
   const Layout = getLayoutByTheme({ theme: siteConfig('THEME'), router: useRouter() })
 
   const router = useRouter()
-  const keyword = getSearchKey(router)
+  const keyword = router?.query?.s
 
   let filteredPosts
   // 静态过滤
@@ -34,15 +32,7 @@ const Search = props => {
     filteredPosts = []
   }
 
-  const meta = {
-    title: `${keyword || ''}${keyword ? ' | ' : ''}${locale.NAV.SEARCH} | ${siteConfig('TITLE')}`,
-    description: siteConfig('DESCRIPTION'),
-    image: siteInfo?.pageCover,
-    slug: 'search',
-    type: 'website'
-  }
-
-  props = { ...props, meta, posts: filteredPosts }
+  props = { ...props, posts: filteredPosts }
 
   return <Layout {...props} />
 }
@@ -61,13 +51,6 @@ export async function getStaticProps() {
     props,
     revalidate: parseInt(BLOG.NEXT_REVALIDATE_SECOND)
   }
-}
-
-function getSearchKey(router) {
-  if (router.query && router.query.s) {
-    return router.query.s
-  }
-  return null
 }
 
 export default Search
