+++ pages/tag/index.js
@@ -1,5 +1,4 @@
 import { getGlobalData } from '@/lib/notion/getNotionData'
-import { useGlobal } from '@/lib/global'
 import BLOG from '@/blog.config'
 import { useRouter } from 'next/router'
 import { getLayoutByTheme } from '@/themes/theme'
@@ -11,21 +10,8 @@ import { siteConfig } from '@/lib/config'
  * @returns
  */
 const TagIndex = props => {
-  const { locale } = useGlobal()
-  const { siteInfo } = props
-
   // 根据页面路径加载不同Layout文件
   const Layout = getLayoutByTheme({ theme: siteConfig('THEME'), router: useRouter() })
-
-  const meta = {
-    title: `${locale.COMMON.TAGS} | ${siteConfig('TITLE')}`,
-    description: siteConfig('DESCRIPTION'),
-    image: siteInfo?.pageCover,
-    slug: 'tag',
-    type: 'website'
-  }
-  props = { ...props, meta }
-
   return <Layout {...props} />
 }
