+++ themes/fukasawa/index.js
@@ -19,7 +19,6 @@ import dynamic from 'next/dynamic'
 import { AdSlot } from '@/components/GoogleAdsense'
 import { Style } from './style'
 import replaceSearchResult from '@/components/Mark'
-import CommonHead from '@/components/CommonHead'
 import { siteConfig } from '@/lib/config'
 import WWAds from '@/components/WWAds'
 
@@ -44,7 +43,7 @@ export const useFukasawaGlobal = () => useContext(ThemeGlobalFukasawa)
  * @constructor
  */
 const LayoutBase = (props) => {
-  const { children, headerSlot, meta } = props
+  const { children, headerSlot } = props
   const leftAreaSlot = <Live2D />
   const { onLoading, fullWidth } = useGlobal()
 
@@ -52,8 +51,6 @@ const LayoutBase = (props) => {
         <ThemeGlobalFukasawa.Provider value={{}}>
 
             <div id='theme-fukasawa'>
-                {/* SEO信息 */}
-                <CommonHead meta={meta}/>
                 <Style/>
 
                 <TopNav {...props} />
