+++ themes/landing/index.js
@@ -15,7 +15,6 @@ import Features from './components/Features'
 import FeaturesBlocks from './components/FeaturesBlocks'
 import Testimonials from './components/Testimonials'
 import Newsletter from './components/Newsletter'
-import CommonHead from '@/components/CommonHead'
 import { useRouter } from 'next/router'
 import CONFIG from './config'
 import Loading from '@/components/Loading'
@@ -30,12 +29,9 @@ import { siteConfig } from '@/lib/config'
  * @returns
  */
 const LayoutBase = (props) => {
-  const { meta, siteInfo, children } = props
+  const { children } = props
 
   return <div id='theme-landing' className="overflow-hidden flex flex-col justify-between bg-white">
-
-        {/* 网页SEO */}
-        <CommonHead meta={meta} siteInfo={siteInfo} />
 
         {/* 顶部导航栏 */}
         <Header />
