+++ themes/matery/index.js
@@ -1,5 +1,4 @@
 import CONFIG from './config'
-import CommonHead from '@/components/CommonHead'
 import TopNav from './components/TopNav'
 import Live2D from '@/components/Live2D'
 import { useGlobal } from '@/lib/global'
@@ -40,7 +39,7 @@ import { siteConfig } from '@/lib/config'
  * @constructor
  */
 const LayoutBase = props => {
-  const { children, meta, siteInfo, post } = props
+  const { children, post } = props
   const { onLoading, fullWidth } = useGlobal()
   const router = useRouter()
   const containerSlot = router.route === '/' ? <Announcement {...props} /> : <BlogListBar {...props} />
@@ -52,8 +51,7 @@ const LayoutBase = props => {
 
   return (
         <div id='theme-matery' className="min-h-screen flex flex-col justify-between bg-hexo-background-gray dark:bg-black w-full">
-            {/* SEO相关 */}
-            <CommonHead meta={meta} siteInfo={siteInfo} />
+
             <Style/>
 
             {/* 顶部导航栏 */}
